# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Front API",
        "version": "0.1"
    },
    "paths": {
...
        "https://api.motovago.com/google-login": {
            "post": {
                "summary": "User Login with Google",
                "description": "Login by google",
                "operationId": "googleLogin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email",
                        "required": true,
                        "example": "example@motovago.com"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name",
                        "required": true,
                        "example": "Jhon Doe"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "example": "1847118732817312"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Jhon Doe"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "5450002211"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "example@motovago.com"
                                        },
                                        "status": {
                                            "type": "tinyint",
                                            "example": "1"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "6f261b5ee8d2517d9c04670abecca99b"
                                        },
                                        "type": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name_en": {
                                                        "type": "string",
                                                        "example": "Standard"
                                                    }
                                                },
                                                "type": "object",
                                                "format": "query"
                                            }
                                        },
                                        "country_guid": {
                                            "type": "string",
                                            "example": "128631-12319283-123123192-213123"
                                        },
                                        "country": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Kuwait"
                                                    }
                                                },
                                                "type": "object",
                                                "format": "query"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "400"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "There is an error occured while logging in please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}